+++ crates/holochain_types/CHANGELOG.md
@@ -7,6 +7,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## \[Unreleased\]
 
+- Added `Default` implementation for `holochain_types::prelude::InstallAppPayload`. You may add `..Default::default` 
+  after other fields if you use this type so that the type can be updated without breaking your code.
+
 ## 0.3.0-beta-dev.13
 
 ## 0.3.0-beta-dev.12